Update avr code from dev branch. Code is not working yet.
The plan is:
The boards avr writes out its pin states to the display pins/header/pads
and it reads the outputs to set from the display header.
This in the loop.
So i can use the commands i already know and which are working.
No interrupts are needed.
I soldered wires to some of the display header/pads on the board.
The display headers will be attached to arduino for easy programming.

	geändert:       software/valve_control.c